refactor(*.*): Fare 객체 패키지 이동
