+++ src/main/java/nextstep/subway/path/domain/Path.java
@@ -3,7 +3,6 @@ package nextstep.subway.path.domain;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import nextstep.subway.line.domain.Fare;
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.Section;
 import nextstep.subway.line.domain.Sections;
