+++ src/main/java/nextstep/subway/path/farePolicy/NoneDiscountPolicy.java
@@ -1,6 +1,6 @@
 package nextstep.subway.path.farePolicy;
 
-import nextstep.subway.line.domain.Fare;
+import nextstep.subway.path.domain.Fare;
 
 public class NoneDiscountPolicy implements MemberDiscountPolicyService{
     @Override
